Removed the email logic from exception handler and added it in collate files, updated version for deployment
